Update climate zone groupby
